requirements.txt
absl-py==2.1.0
ale-py==0.8.1
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.7.0
eventlet==0.37.0
Farama-Notifications==0.0.4
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
fonttools==4.54.1
fpdf2==2.8.1
fsspec==2024.9.0
greenlet==3.1.1
grpcio==1.66.2
gymnasium==0.29.1
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mininet==2.3.0.dev6
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.28.2
psutil==6.0.0
pycparser==2.22
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.1.4
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.1
scapy==2.6.0
scikit-learn==1.5.2
scipy==1.14.1
Shimmy==1.3.0
simple-websocket==1.1.0
six==1.16.0
stable_baselines3==2.3.2
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
